Only check records that have changed since the last update.
Drupal(changed) is the same as OSV(modified).

Get the highest 'modified' date from the OSV records.
Fetch from the api, sorted by changed and ordered so newer changed items
are first.
Only process records with a 'changed' value higher than the highest 'modified'.
If a record is not newer than the highest modified we set 'fetch_again'
to False. If this is not hit it will fetch another page of results.