feat: split script into a "download" script and a "generate" script (#5)
The current script is a single massive Python file with a number of
functions for generating OSVs from alternative sources such as the RSS
feed which we've decided not to use for now, and it also does both the
downloading and generating as an interweaved process which can make it
hard to iterate on since most of the time we'll want to be regenerating
all the OSVs in response to a fix or improvement without hammering the
same external sources everytime.

To make it easier to iterate, I've broken the single script into one
focused on downloading the SA advisories and another focused on the
generating of the OSVs from said advisories - the bulk of the logic is
still in the latter script since it turns out that's actually where most
of our API calls are, but this is still a good step towards making
generating OSVs more robust.

I've also introduced some types as part of this which I'll expand as
part of followup work.

I have done some basic updating of the readme, but purposely not made
much of an attempt to ensure the docs are aligned with the state of the
scripts (in particular in regards to the alternative sources) as I think
that'll be a lot easier to do once we've made more progress through the
improvements.