feat: build out types further (#6)
This builds out our types further especially for the OSV side of things
for which I've mirrored the
[structs](https://github.com/ossf/osv-schema/blob/main/bindings/go/osvschema/vulnerability.go)
and
[constants](https://github.com/ossf/osv-schema/blob/main/bindings/go/osvschema/constants.go)
from our Go bindings.

I'm pretty sure the types are not actually completely correct mainly in
regard to if a property can be present vs `None`, but that should be
very easy to iterate on as we use them.

Also since these are quite large now I've attempted to split them up
into their own files/namespaces - I'm not sure if I've done it in the
most pythonic way, but it seems to work and is cleaner than having it
all in one file so I think we should just ship it for now 🤷

Related, I've still not really managed to figure out the rules of the
module system either which is why the directory is called `typings`
(which isn't actually a terrible name), as PyCharm seemed to think I
could do `import scripts.types` but Python would error about there not
being a module named `scripts` and then if I removed the `scripts.` then
Python assumed I wanted the built-in `types` module.