feat: optimize node entity requests and caching (#7)
This has us use the dedicated node api endpoint for retrieving data on
specific nodes rather than using the nodes list endpoint, as the former
is both more semantically accurate and should in theory be more
performant as it should be more likely that the cache is hot.

We also now optimistically assume the node is already cached on disk as
that is more likely (especially if we introduce precaching leveraging
the list endpoint), which saves us an extra system call and technically
accounts for the file being removed in the time after the existence
check but before the read (though it was very unlikely that would ever
happen).

The cache itself has been moved from `files` to `cache/nodes` and no
longer includes the type as we cannot accurately determine that from the
advisory alone, and while most of the nodes are either `project_module`
or `project_release` there is at least one `project_core` and
`project_theme` type nodes; this is also why I've renamed
`drupal.ProjectModule` to `drupal.Project` and inlined the dedicated
node type fetching functions.