refactor: improve type safety of `drupal.ApiResponse` (#21)
It seems in Python generics default to `Any` even if they have an upper
bound, so we should explicitly be defining `Node` as the default type
for `drupal.ApiResponse`.

I've also switched us to using the new type parameter syntax introduced
in [PEP-695](https://peps.python.org/pep-0695/) and available in Python
3.13, which means we no longer have to define a separate variable for
the generic.