refactor: be strict with types (#26)
While we are currently type checking our code with `mypy`, we're doing
so in a loose manner meaning it's still possible for bugs and
hard-to-maintain code to slip through - this enables `strict` mode to
help address that, and addresses the current violations in our code.

I have had to use an `Any` for the parameter to the `semver_for_sorting`
function as it is assigning integer values to what is originally a
string, and so needs a more dedicated rewrite than what I'd like to
cover in this change.

You can check out the individual commits to see what kind of changes
each `strict` flag required