feat: don't process the "fixed in" field (#29)
Currently we're generating the affected events using data from both the
`field_affected_versions` (which is a string with one or more composer
version constraints/ranges) and `field_fixed_in` (which holds an array
of references to project releases which fix the advisory) which could be
useful except right now the fields frequently contradict each other - I
suspect that part of the reason for this is that the
`field_affected_versions` is for Drupal 8+ versions whereas
`field_fixed_in` includes Drupal 7 releases, but there's also a number
that are just plain wrong.

While this will be useful in helping ensure advisories have the correct
information, right now it's impeding work on restructuring the
`affected` ranges on our advisories which is a more pressing matter so
this patch removes our usage of the `field_fixed_in` field for the time
being.

Once we have gotten the `affected` ranges correctly (re)structured and
the parsing of `field_affected_versions` refined, we'll be in a much
better position to revisit `field_fixed_in`