feat: generate a dedicated range for each version constraint (#33)
Since each range should only have one `introduced` and either a `fixed`
or `last_affected` event, we should be generating a range for every
version constraint in the version range provided by an advisory.

Note that this actually causes us to have a few invalid advisories
because the sorting wipes away invalid `introduced` events which are
currently being generated as a result of flaws in our version constraint
parser which will be fixed shortly.

For now I've managed to leverage the existing sorting code to do
something best-effort, but it'll be greatly cleaned up once we start
only generating valid semantic versions which is why I've used such a
horrible function name and have just grabbed the first event without
bothering to actually find the `introduced` (though, that is what the
first event should be)