feat: rewrite affected version parser to match `composer` and report warnings about invalid versions (#35)
This overhauls our version constraints parser to ensure we're performing
almost in accordance with
[composer](https://getcomposer.org/doc/articles/versions.md) and with
support for surfacing warnings so that we can identify invalid versions
to get addressed upstream.

The primary difference in how we parse versions compared with `composer`
is that we assume a lack of stability constraint means `stable` in all
cases, instead of being [dependent on the
operator](https://getcomposer.org/doc/articles/versions.md#stability-constraints)
- ideally, the stability should be included as part of the constraint
which might be something we can convince the upstream to do, but for now
I've chosen to differ for a few key reasons:
1. Drupal security advisories generally only cover stable versions, so
it's reasonable to assume whoever wrote the version constraints for a
particular advisory was only thinking about stable versions
2. Being semantically correct would mean a _lot_ of changes to our
advisories
3. Ideally we want to be referring versions that actually exist, which
is most likely the case for the versions explicitly mentioned in
particular version constraints (in part because of point 1.), whereas if
we follow the spec we'd have to assume a lot of the time that there's a
`-dev` version that probably won't exist

I have done the implementation to first be spec-compliant, and then
modified it to reflect the above, meaning we should be able to easily
change back in future if needed by reviewing the last handful of commits
on this branch.

I've not actually wired anything up to use the warnings returned by the
parser given the size of this patch already, but they are asserted by
our new test suite.

There's definitely some cleanup worth doing with the implementation, in
particular I probably should have named the component properties
`major`, `minor`, and `patch`, but I'm going to try tackle most of those
as a follow-up since everything is working and renaming things might
make it harder to review the individual commits, and this should be
stable enough to let us progress to the next phase where we'll be
engaging with the Drupal community to improve the advisory data which
should feed back heavily into how robust this implementation actually
needs to be