feat: capture any warnings as part of the advisories (#46)
This will make it easier to determine when an advisory needs improving,
and to confirm when a "patch" has successfully addressed a warning,
without bloating all our advisories since the majority of them will be
warning free.

Note that I've already got fixes for the advisories with warnings in
this patch, I just want to have the warnings captured in a "before"
commit so that my follow-up patch will be ~more impressive~ show the
warnings are correctly addressed.