ci: create workflow for creating pull requests updating advisories (#49)
This works by running through the scripts like we do locally, and then
using the `gr2m/create-or-update-pull-request-action` action to commit
the results to a particular branch, creating or updating a pull request
if there are changes.

In order to have CI run for this pull request, the workflow also
manually triggers a workflow run on its own pull request using the
GitHub API.

Note that I've used this setup in the past on other projects, and know
@gr2m as we both help maintain some GitHub npm packages so I have high
trust in these actions.

For now I've just set this up to be triggerable via `workflow_dispatch`,
but the goal is to later have it run on a schedule - I've also not setup
any caching as this usually takes less than 5 minutes to go through
without any cache so I think we should be good to add that as a
follow-up