feat: prefix advisory IDs with `D` instead of `OSV-` and retain uppercasing (#68)
The OSV schema requires advisories to use an id that is prefixed with an
identifier that represents their "home database" - we're currently using
`OSV-` so that our advisories are valid when compared to the JSON
schema, but that's not really correct since that's a different database.

This switches us to use `DSA` (for "Drupal Security Advisory") as that
is only one letter different from the `SA` prefix Drupal already uses
(which we can't use as its too generic) and it happens to already be a
defined prefix in the JSON schema as it used by the Debian community
(for "Debian Security Advisory") - while this overlap isn't ideal, it
should be fine for now at least since the databases are for very
different ecosystems and Drupal advisory IDs use `-CORE` and `-CONTRIB`
after their prefix so there's very little likelihood of a ID collision.

I have already confirmed with Google this should be acceptable for now,
but it's still possible we'll change this once we talk more with the
Drupal security team as ideally it would be best if everyone uses the
same ID including Drupal themselves - one possible solution that
wouldn't require a change would be to make the `-CORE` and `-CONTRIB`
part of the prefix which is valid though does make it a little long.