ci: replace pull request creator action (#69)
The current workflow doesn't seem to be recreating the branch how I'd
expect, e.g. the current pull request is conflicted due to #68. I think
this is a result of how the action is implemented, and that switching to
`peter-evans/create-pull-request` will resolve this as it looks like
that uses a different implementation aimed to account for this kind of
situation.

This action is also more popular and what we're using in other codebases