feat: "patch" advisories whose affected versions ranges have warnings (#52)
Being able to "patch" advisories ourselves lets us verify changes before
submitting them upstream and also roll fixes to the data out faster
since we don't need to wait for the upstream to (re)publish the changes
(which would likely need to go through their own review process first).

The actual implementation is pretty straightforward and doesn't even
require any additional dependencies since Python 3.11 added `tomllib` as
a standard library (which I've used instead of JSON as it lets us have
comments more naturally).

See the new section of the readme for more details on how this works and
why the file is structured as it is