feat: retry request if api returns a 429 when precaching nodes (#90)
We've had a couple of our (re)generation runs fail when precaching the
Drupal nodes due to 429s - we might as well include a basic "retry once"
attempt to help avoid this. While it's possible this could happen for
any API request it should be very unlikely except for this one which is
why I've not changed them.

In the long-run if this becomes more regular (or requested by the Drupal
team) we'll probably move to a more dedicated HTTP client implementation
to handle this across the board