chore: add colour to the advisory generation output. (#109)
I find the output from the Advisory generation to be an unreadable wall
of text.

This adds a scattering of colour to the notices for SAs that have had an
action applied to them.

This is manage through the addition of
[Colorama](https://pypi.org/project/colorama/). I selected this one as
it manages the colours and styles while taking into account the platform
it's run on, is actively maintained, and has been around for more than a
decade.

<img width="1125" height="1072" alt="image"
src="https://github.com/user-attachments/assets/90966c1a-3884-448a-b302-f831648b2e3c"
/>

I note that the `poetry.lock` has an unexpected removal. Could this be
because it is included via the new Colourama?